create inheritance between cards to avoid repetition
