+++ lib/views/course_view/downloaded_courses_view/download_content_view.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:gocast_mobile/views/components/base_view.dart';
 import 'package:gocast_mobile/views/components/custom_search_top_nav_bar.dart';
-import 'package:gocast_mobile/views/course_view/downloaded_courses_view/download_card.dart';
+import 'package:gocast_mobile/views/course_view/components/download_card.dart';
 
 /// CourseListScreen
 ///
